backend/requirements.txt
Django==3.2.8
django_on_heroku==1.1.2
djangorestframework==3.12.4
drf_yasg==1.20.0
Pillow==8.4.0
requests==2.26.0
typing_extensions==3.10.0.0
